% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards_manip.R
\name{create_board}
\alias{create_board}
\alias{delete_board}
\alias{edit_board}
\title{Create, delete, or edit a Board}
\usage{
create_board(name, description, token)

delete_board(board, token)

edit_board(board, name = NULL, description = NULL, token)
}
\arguments{
\item{name}{name of the new board for creation, new name to give for edit.}

\item{description}{description of the board?}

\item{token}{a character string with an access token generated with `pinterest_token()`}

\item{board}{a board name.}
}
\value{
Either a data.frame, or an error.
}
\description{
Create, delete, or edit a Board
}
\examples{
\dontrun{
create_board("testboard", "from rpinterest", token)
delete_board("colinfay/testboard", token)
edit_board("colinfay/testboard",description = "A new desc", token = token)
}
}
