% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_pins.R
\name{get_board_pins_by_id}
\alias{get_board_pins_by_id}
\alias{get_board_pins_by_name}
\title{Retrieve the Pins on a Board}
\usage{
get_board_pins_by_id(id, token)

get_board_pins_by_name(user, board, token)
}
\arguments{
\item{id}{a character string with an ID.}

\item{token}{a character string with an access token generated with `pinterest_token()`}

\item{user}{a character string with a user name.}

\item{board}{a character string with a board name.}
}
\value{
Either a data.frame, or an error.
}
\description{
Get information about all the pins on a pinterest board using the board ID or user/name.
}
\examples{
\dontrun{
get_board_pins_by_id(id = "42080646457333782", token = token)
get_board_pins_by_name(user = "colinfay", board = "violet-mon-amour", token = token)
}
}
