% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoardPinsByName.R
\name{BoardPinsByName}
\alias{BoardPinsByName}
\title{Get board pins using board name}
\usage{
BoardPinsByName(user, board, token)
}
\arguments{
\item{user}{a character string with a user name.}

\item{board}{a character string with a board name.}

\item{token}{a character string with an access token generated at \url{https://developers.pinterest.com/tools/access_token/}}
}
\value{
Always returns a data.frame, with a warning when appropriate.
}
\description{
Get information about all the pins on a pinterest board using the board name.
}
\details{
Takes a user name, a board name and an access token, returns a data.frame.
}
\examples{
BoardPinsByName(user = "colinfay", board = "blanc-mon-amour", token = "your_token")
}

