% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoardSpecByID.R
\name{BoardSpecByID}
\alias{BoardSpecByID}
\title{Get board using its ID}
\usage{
BoardSpecByID(boardID, token)
}
\arguments{
\item{boardID}{a character string with a board ID.}

\item{token}{a character string with an access token generated at \url{https://developers.pinterest.com/tools/access_token/}}
}
\value{
Always returns a data.frame, with a warning when appropriate.
}
\description{
Get information about a pinterest board using the board ID.
}
\details{
Takes a board ID and an access token, returns a data.frame.
}
\examples{
BoardSpecByID(boardID = "42080646457333782", token = "your_token")
}

