% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoardSpecByName.R
\name{BoardSpecByName}
\alias{BoardSpecByName}
\title{Get board using its name}
\usage{
BoardSpecByName(user, board, token)
}
\arguments{
\item{user}{a character string with an user name.}

\item{board}{a character string with a board name.}

\item{token}{a character string with an access token generated at \url{https://developers.pinterest.com/tools/access_token/}}
}
\value{
Returns a data.frame with board specifications.
}
\description{
Get informations about a pinterest board using its name.
}
\details{
Takes a user name, a board name and an access token.
}
\examples{
BoardSpecByName(user = "colinfay", board = "blanc-mon-amour", token = token)
}

