% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_spec.R
\name{get_board_spec_by_name}
\alias{get_board_spec_by_name}
\title{Get board pins using board name}
\usage{
get_board_spec_by_name(user, board, token)
}
\arguments{
\item{user}{a character string with a user name.}

\item{board}{a character string with a board name.}

\item{token}{a character string with an access token generated with `pinterest_token()`}
}
\value{
Always returns a data.frame, with a warning when appropriate.
}
\description{
Get information about all the pins on a pinterest board using the board name.
}
\details{
Takes a user name, a board name and an access token, returns a data.frame.
}
\examples{
\dontrun{
get_board_spec_by_name(user = "colinfay", board = "blanc-mon-amour", token = token)
}
}
