% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{get_logged_user}
\alias{get_logged_user}
\alias{get_logged_user_boards}
\alias{get_logged_user_boards_suggestions}
\title{Get Information about the logged user}
\usage{
get_logged_user(token)

get_logged_user_boards(token)

get_logged_user_boards_suggestions(count = 25, token)
}
\arguments{
\item{token}{a character string with an access token generated with `pinterest_token()`}

\item{count}{for suggestions, the max number of suggestion to retrieve.}
}
\value{
Either a data.frame, or an error.
}
\description{
Get user information about their account or about
their boards.
}
\examples{
\dontrun{
get_logged_user(token)
get_logged_user_boards(token)
get_logged_user_boards_suggestions(token = token)
}
}
