% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_spec.R
\name{get_pin_spec_by_id}
\alias{get_pin_spec_by_id}
\title{Return information about a Pin}
\usage{
get_pin_spec_by_id(id, token)
}
\arguments{
\item{id}{a character string with a pin ID.}

\item{token}{a character string with an access token generated with `pinterest_token()`}
}
\value{
Either a data.frame, or an error.
}
\description{
Get information about a pinterest pin using the pin ID.
}
\examples{
\dontrun{
get_pin_spec_by_id(id = "42080577745042298", token = token)
}

}
