% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_spec.R
\name{get_user_spec_by_id}
\alias{get_user_spec_by_id}
\alias{get_user_spec_by_name}
\title{Return a user's information}
\usage{
get_user_spec_by_id(id, token)

get_user_spec_by_name(user, token)
}
\arguments{
\item{id}{a character string with a user ID.}

\item{token}{a character string with an access token generated with `pinterest_token()`}

\item{user}{a character string with a user name.}
}
\value{
Either a data.frame, or an error.
}
\description{
Get information about a Pinterest user using the ID or name.
}
\examples{
\dontrun{
get_user_spec_by_id(id = "42080715176677612", token = token)
get_user_spec_by_name(user = "colinfay", token = token)
}
}
