% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_spec.R
\name{get_user_spec_by_id}
\alias{get_user_spec_by_id}
\title{Get a user using its ID}
\usage{
get_user_spec_by_id(id, token)
}
\arguments{
\item{id}{a character string with a user ID.}

\item{token}{a character string with an access token generated with `pinterest_token()`}
}
\value{
Always returns a data.frame, with a warning when appropriate.
}
\description{
Get information about a pinterest user using the user ID.
}
\details{
Takes a user ID and an access token, returns a data.frame.
}
\examples{
\dontrun{
get_user_spec_by_id(id = "42080715176677612", token = token)
}
}
