% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_to_pinterest.R
\name{ggplot_to_pinterest}
\alias{ggplot_to_pinterest}
\title{Send a ggplot plot to a Pinterest board}
\usage{
ggplot_to_pinterest(plot, board, note, link = NULL, token, ...)
}
\arguments{
\item{plot}{a ggplot object}

\item{board}{the name of the board}

\item{note}{description of the image}

\item{link}{a link to put under the image}

\item{token}{a character string with an access token generated with `pinterest_token()`}

\item{...}{arguments to be passed to `ggsave()`}
}
\value{
Either a data.frame, or an error.
}
\description{
Send a ggplot plot to a Pinterest board
}
\examples{
\dontrun{
library(ggplot2)
x <- ggplot(iris, aes(Sepal.Length, Sepal.Width)) + geom_point()
ggplot_to_pinterest(x, "colinfay/gris-mon-amour", "test rpinterest", token = token)
}

}
