% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_manip.R
\name{create_pin}
\alias{create_pin}
\alias{delete_pin}
\alias{edit_pin}
\title{Create, delete, or edit a Pin}
\usage{
create_pin(file, board, note, link = NULL, token)

delete_pin(id, token)

edit_pin(id, board = NULL, note = NULL, link = NULL, token)
}
\arguments{
\item{file}{the path to the file}

\item{board}{name of the board (required)}

\item{note}{a description (required)}

\item{link}{an url (optional)}

\item{token}{a character string with an access token generated with `pinterest_token()`}

\item{id}{a character string with an ID.}
}
\value{
Either a data.frame, or an error.
}
\description{
Create, delete, or edit a Pin
}
\examples{
\dontrun{
create_pin("~/Pictures/graph.png", "my_board", "a graph", token = token)
delete_pin("42080577756375754", token)
edit_pin("42080577744990466",note = "run run run", token = token)
}
}
