% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PinSpecByID.R
\name{PinSpecByID}
\alias{PinSpecByID}
\title{Get pin using its ID}
\usage{
PinSpecByID(pinID, token)
}
\arguments{
\item{pinID}{a character string with a pin ID.}

\item{token}{a character string with an access token generated at \url{https://developers.pinterest.com/tools/access_token/}}
}
\value{
Always returns a data.frame, with a warning when appropriate.
}
\description{
Get information about a pinterest pin using the pin ID.
}
\details{
Takes a pin ID and an access token, returns a data.frame.
}
\examples{
PinSpecByID(pinID = "42080577745042298", token = "your_token")
}

