% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{pinterest_token}
\alias{pinterest_token}
\title{Create a Pinterest API token}
\usage{
pinterest_token(app, app_id, app_secret, verbose = TRUE,
  callback = "https://colinfay.me/rpinterestcallback/")
}
\arguments{
\item{app}{the name of your app}

\item{app_id}{the ID of your app}

\item{app_secret}{the secret of the app}

\item{verbose}{wether of not to print a message about the auth process}

\item{callback}{the callback uri for your app}
}
\value{
an access token
}
\description{
Create a Pinterest API token
}
