% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rpinterest-deprecated}
\alias{rpinterest-deprecated}
\alias{BoardPinsByID}
\alias{BoardPinsByName}
\alias{BoardSpecByID}
\alias{BoardSpecByName}
\alias{PinSpecByID}
\alias{UserSpecByID}
\alias{UserSpecByName}
\title{Deprecated functions from \pkg{rpinterest}.}
\usage{
BoardPinsByID(boardID, token)

BoardPinsByName(user, board, token)

BoardSpecByID(boardID, token)

BoardSpecByName(user, board, token)

PinSpecByID(pinID, token)

UserSpecByID(userID, token)

UserSpecByName(user, token)
}
\description{
Deprecated functions from \pkg{rpinterest}.
}
\keyword{internal}
