% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserSpecByID.R
\name{UserSpecByID}
\alias{UserSpecByID}
\title{Get a user using its ID}
\usage{
UserSpecByID(userID, token)
}
\arguments{
\item{userID}{a character string with a user ID.}

\item{token}{a character string with an access token generated at \url{https://developers.pinterest.com/tools/access_token/}}
}
\value{
Returns a data.frame with user informations.
}
\description{
Get informations about a pinterest user using its ID.
}
\details{
Takes a user ID and an access token, returns a data.frame.
}
\examples{
UserSpecByID(user = "42080715176677612", token = token)
}

