% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{get_logged_user}
\alias{get_logged_user}
\alias{get_logged_user_boards}
\title{Get Information about the logged user}
\usage{
get_logged_user(token)

get_logged_user_boards(token)
}
\arguments{
\item{token}{a character string with an access token generated with `pinterest_token()`}
}
\value{
a data.frame
}
\description{
Get user information about their account or about
their boards.
}
\examples{
\dontrun{
get_logged_user(token)
get_logged_user_boards(token)
}
}
