% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_spec.R
\name{get_pin_spec_by_id}
\alias{get_pin_spec_by_id}
\title{Get pin using its ID}
\usage{
get_pin_spec_by_id(id, token)
}
\arguments{
\item{id}{a character string with a board ID.}

\item{token}{a character string with an access token generated with `pinterest_token()`}

\item{pinID}{a character string with a pin ID.}
}
\value{
Always returns a data.frame, with a warning when appropriate.
}
\description{
Get information about a pinterest pin using the pin ID.
}
\details{
Takes a pin ID and an access token, returns a data.frame.
}
\examples{
get_pin_spec_by_id(id = "42080577745042298", token = token)
}
